% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/add_row.R
\name{add_row}
\alias{add_row}
\title{add_col}
\usage{
add_row(original, index, new_object)
}
\arguments{
\item{original}{Initial variable where the final matrix will be stored, it can be already matrix.}

\item{new_object}{Row or matrix to be inserted in the original one.}

\item{row_index}{Numeric index representing the row where the new line should be inserted.}
}
\value{
The matrix with the changes done after the function was run.
}
\description{
Function that adds new rows into a matrix/vector (depending the case).
It works simulating the behavior of a matrix in Matlab.
Everytime you want to add a row or a full matrix in an out of bounds index, automatically is created with the row given.
It internally calls the add_col function to do this.
}

