% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate}
\alias{evaluate}
\title{evaluate}
\usage{
evaluate(nmmso_state, chg, problem_function)
}
\arguments{
\item{nmmso_state}{Structure holding state of swarm.}

\item{chg}{}

\item{problem_function}{String containing name of function to be optimised.}

\item{problem_function_params}{Meta-parameters needed by problem function.}
}
\value{
nmmso_state = Structure holding state of swarm.
mode_shift = Numeric flag that determines the number of evaluations done for future locations.
y = Result of evaluating the problem_function.
}
\description{
evaluate
}

