% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/evaluate_first.R
\name{evaluate_first}
\alias{evaluate_first}
\title{evaluate_first}
\usage{
evaluate_first(swarm, problem_function, nmmso_state, swarm_size, mn, mx)
}
\arguments{
\item{swarm}{Swarm object}

\item{problem_function}{String containing name of function to be optimised.}

\item{nmmso_state}{Structure holding state of swarm.}

\item{swarm_size}{Maximum number of elements (particles) per swarm.}

\item{mn}{Minimum design parameter values (a vector with param_num elements).}

\item{mx}{Maximum design parameter values (a vector with param_num elements).}

\item{problem_function_params}{Meta-parameters needed by problem function.}
}
\value{
swarm = Swarm with new locations and values.
nmmso_state = Structure with the new locations after evaluation.
}
\description{
Calculates the first evaluation of the swarms.
}

