% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_mid.R
\name{evaluate_mid}
\alias{evaluate_mid}
\title{evaluate_mid}
\usage{
evaluate_mid(nmmso_state, chg, problem_function)
}
\arguments{
\item{nmmso_state}{Structure holding state of swarm.}

\item{chg}{current index}

\item{problem_function}{String containing name of function to be optimised.}

\item{test_function_params}{Meta-parameters needed by problem function.}
}
\value{
nmmso_state = Structure holding state of swarm.
mode_shift = Numeric flag that determines the number of evaluations done for future locations.
y = Result of evaluating the problem_function.
}

