% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_new_locations.R
\name{evaluate_new_locations}
\alias{evaluate_new_locations}
\title{evaluate_new_locations}
\usage{
evaluate_new_locations(nmmso_state, problem_function, I)
}
\arguments{
\item{nmmso_state}{Structure holding state of swarm.}

\item{I}{Vector of indixes for new locations.}

\item{problem_function_params}{Meta-parameters needed by problem function.}
}
\value{
nmmso_state = Structure holding state of swarm.
number_of_new_location = The length of the I vector.
}

