% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_modes.R
\name{extract_modes}
\alias{extract_modes}
\title{extract_modes}
\usage{
extract_modes(nmmso_state)
}
\arguments{
\item{nmmso_state}{Structure holding state of swarm.}
}
\value{
RES = Modes from the given nmmso_state for the X-axis.
RES_Y = Modes from the given nmmso_state for the Y-axis.
}
\description{
Get the mode estimate locations.
}

