% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_initial_locations.R
\name{get_initial_locations}
\alias{get_initial_locations}
\title{get_initial_locations}
\usage{
get_initial_locations(nmmso_state, mn, mx)
}
\arguments{
\item{nmmso_state}{Structure holding state of swarm.}

\item{mn}{Minimum design parameter values (a vector with param_num elements).}

\item{mx}{Maximum design parameter values (a vector with param_num elements).}
}
\value{
nmmso_state = Structure with the new locations.
}
\description{
Calculates the initial locations of the algorithm.
}

