% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hive.R
\name{hive}
\alias{hive}
\title{hive}
\usage{
hive(nmmso_state, problem_function, mn, mx, max_evol, swarm_size)
}
\arguments{
\item{nmmso_state}{Structure holding state of swarm.}

\item{problem_function}{String containing name of function to be optimised.}

\item{mn}{Minimum design parameter values (a vector with param_num elements).}

\item{mx}{Maximum design parameter values (a vector with param_num elements).}

\item{max_evol}{Maximum number of swarms to update in a generation. If not provided this is set at 100.}

\item{swarm_size}{Maximum number of elements (particles) per swarm.}

\item{problem_function_params}{Meta-parameters needed by problem function.}
}
\value{
nmmso_state = Structure holding state of swarm after calculations.
number_of_new_samples = Number of samples created after hive.
}

