% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/increment_swarm.R
\name{increment_swarm}
\alias{increment_swarm}
\title{increment_swarm}
\usage{
increment_swarm(nmmso_state, index, mn, mx, swarm_size)
}
\arguments{
\item{nmmso_state}{Structure holding state of swarm.}

\item{index}{index of the current}

\item{mn}{Minimum design parameter values (a vector with param_num elements).}

\item{swarm_size}{Maximum number of elements (particles) per swarm.}
}
\value{
nmmso_state = Structure holding state of swarm.
}
\description{
TODO: This need to be documented extensively
}

