% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/merge_swarms.R
\name{merge_swarms}
\alias{merge_swarms}
\title{merge_swarms}
\usage{
merge_swarms(nmmso_state, problem_function, mn, mx)
}
\arguments{
\item{nmmso_state}{Structure holding state of swarm.}

\item{problem_function}{String containing name of function to be optimised.}

\item{mn}{Minimum design parameter values (a vector with param_num elements).}

\item{mx}{Maximum design parameter values (a vector with param_num elements).}

\item{problem_function_params}{Meta-parameters needed by problem function.}
}
\value{
nmmso_state = Structure holding state of swarm.
number_of_mid_evals = Number of evaluations done for the merging.
}
\description{
merge_swarms
}

