% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/merge_swarms_together.R
\name{merge_swarms_together}
\alias{merge_swarms_together}
\title{merge_swarms_together}
\usage{
merge_swarms_together(swarm1, swarm2)
}
\arguments{
\item{swarm1}{The first swarm.}

\item{swarm2}{The second swarm.}
}
\value{
The joined swarm
}
\description{
Joins two swarms into one.
}

