% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nmmso_iterative.R
\name{NMMSO_iterative}
\alias{NMMSO_iterative}
\title{NMMSO_iterative}
\usage{
NMMSO_iterative(swarm_size, problem_function, max_evaluations, mn, mx,
  evaluations, nmmso_state, max_evol = 100, tolerance_value = (10^-6),
  c_1 = 2, c_2 = 2, omega = 0.1)
}
\arguments{
\item{swarm_size}{Maximum number of elements (particles) per swarm.}

\item{problem_function}{String containing name of function to be optimised.}

\item{max_evaluations}{Maximum number of evaluations to be taken through the problem function.}

\item{mn}{Minimum design parameter values (a vector with param_num elements).}

\item{mx}{Maximum design parameter values (a vector with param_num elements).}

\item{evaluations}{Number of evaluations expended already. If set at 0, this will initialise the swarm_state structure and run the
algorithm for a single generation, otherwise it will run it for a single generation from the evaluations number inputted.}

\item{nmmso_state}{Structure holding state of swarm. Can be empty or omitted if evaluations set at zero, must be provided if evals > 0.}

\item{max_evol}{Maximum number of swarms to update in a generation. If not provided this is set at 100.}

\item{tolerance_value}{Tolerance value for merging automatically (default 10^-6).}

\item{c_1}{constraint 1 on velocity (default 2.0)}

\item{c_2}{constraint 2 on velocity (default 2.0)}

\item{omega}{inertia of a particle (default 0.1)}
}
\value{
mode_loc = Design space location of current mode estimates (swarm gbests), note that at least one is likely to be very poor due to the
new swarm spawning at the end of each generation, and that these will be a combination of both global and local mode estimate.
mode_y = Function evalutions corresponding to the mode estimates.
evaluations = Number of problem function evaluations until this point.
nmmso_state = Structure holding the state of the swarms. Unless you want to pick apart the details of how the algorithm searchs the space,
then the only two elements you will probably be interested in are X and Y which are preallocated matrices to hold all locations visited
(therefore nmmso_state.X(1:evaluations,:) will hold all the design space locations visited by the optimiser thus far.
}
\description{
Implementation of the Niching Migratory Multi-Swarm Optimser.
}

