% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_new.R
\name{random_new}
\alias{random_new}
\title{random_new}
\usage{
random_new(nmmso_state, problem_function, mn, mx, swarm_size)
}
\arguments{
\item{nmmso_state}{Structure holding state of swarm.}

\item{problem_function}{String containing name of function to be optimised.}

\item{mn}{Minimum design parameter values (a vector with param_num elements).}

\item{mx}{Maximum design parameter values (a vector with param_num elements).}

\item{swarm_size}{Maximum number of elements (particles) per swarm.}

\item{problem_function_params}{Meta-parameters needed by problem function.}
}
\value{
nmmso_state = Structure holding state of swarm.
number_rand_modes = Number of random modes (Apparently always 1).
}

