% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/UNI.R
\name{UNI}
\alias{UNI}
\title{UNI}
\usage{
UNI(x1, x2)
}
\arguments{
\item{x1}{Vector number 1}

\item{x2}{Vector number 2}
}
\value{
x_c = x1 crossovered
x_d = x2 crossovered
}
\description{
Simulates a binary crossover.
}

