% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniform_sphere_points.R
\name{uniform_sphere_points}
\alias{uniform_sphere_points}
\title{uniform_sphere_points}
\usage{
uniform_sphere_points(n, d)
}
\arguments{
\item{n}{number of points}

\item{d}{dimension}
}
\value{
X = the matrix generated#'
}
\description{
Function that generates n points uniformly within the unit sphere in d dimensions.#'
}

