% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/increment_swarm.R
\name{increment_swarm}
\alias{increment_swarm}
\title{increment_swarm}
\usage{
increment_swarm(nmmso_state, index, mn, mx, swarm_size, c_1, c_2, omega)
}
\arguments{
\item{nmmso_state}{Structure holding state of swarm.}

\item{index}{index of the current}

\item{mn}{Minimum design parameter values (a vector with param_num elements).}

\item{swarm_size}{Maximum number of elements (particles) per swarm.}

\item{c_1}{constraint 1 on velocity}

\item{c_2}{constraint 2 on velocity}

\item{omega}{inertia of a particle}
}
\value{
nmmso_state = Structure holding state of swarm.
}
\description{
increment_swarm
}

